# Conflicts:
#	src/main/java/com/social/demo/dao/repository/IStudentService.java

+++ src/main/java/com/social/demo/entity/Ranking.java
@@ -11,7 +11,7 @@ public class Ranking {
      /**
      *学生成绩
      */
-     private Integer score;
+     private Double score;
     /**
      * 学生各科
      */
